requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
openai>=1.40.0
redis>=5.0.0
chromadb>=0.5.0
tiktoken>=0.7.0
python-dotenv>=1.0.0
pydantic>=2.0.0
